added missing integer cast at return value

+++ windows/hid.c
@@ -1351,7 +1351,7 @@ int HID_API_EXPORT HID_API_CALL hid_send_output_report(hid_device* dev, const un
 		return -1;
 	}
 
-	return length;
+	return (int) length;
 }
 
 int HID_API_EXPORT HID_API_CALL hid_get_input_report(hid_device *dev, unsigned char *data, size_t length)
